backend/requirements.txt
Flask==3.0.3
flask-smorest==0.44.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
alembic==1.13.0
pydantic==2.5.0
pydantic[email]==2.5.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
marshmallow==3.20.1
apispec==6.3.1
Flask-Cors==4.0.0
